pyproject.toml
[tool.poetry]
name = "envoy_extproc_sdk"
version = "0.1.3"
description = ""
authors = ["W. Ross Morrow <morrowwr@gmail.com>"]
packages = [
    { include = "envoy_extproc_sdk" },
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.46.1"
protoc-gen-validate = "^0.4.2"
grpclib = "^0.4.2"
ddtrace = "^1.1.*"
datadog = "^0.44.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^3.9.2"
isort = "^5.9.3"
mypy = "^0.910"
pytest = "^7.1.2"
mypy-protobuf = "^3.2.0"
pytest-asyncio = "^0.18.3"

[tool.black]
line-length = 100
include = 'envoy_extproc_sdk\/.*\.py$|examples\/.*\.py$|tests\/.*\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \venv
    | _build
    | buck-out
    | build
    | dist
    | generated
  )/
)
'''

[tool.isort]
src_paths = ["config", "tests"]
profile = "black"
line_length = 80
multi_line_output = 3
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
force_sort_within_sections = true
force_alphabetical_sort_within_sections = true
extend_skip = "generated"
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
